ai-service/requirements.txt
flask==2.3.3
rembg==2.0.50
onnxruntime==1.16.3
pillow==10.1.0
gunicorn==21.2.0